+++ src/tests/PagedListResultNet5Tests/GetPagedWithFiltersTests.cs
@@ -343,7 +343,11 @@ namespace PagedListResultNet5Tests
                         FilterApplyOrder = 1
                     }
                 },
-                PredefinedRecords = new List<string> { "5" }
+                //PredefinedRecords = new List<string> { "5" }
+                PredefinedRecord = new DataPredefinedFilterDefinition()
+                {
+                    PredefinedRecords = new List<string> { "5" }
+                }
             };
 
             var query = _dbContext.Posts
@@ -417,7 +421,11 @@ namespace PagedListResultNet5Tests
                         FilterApplyOrder = 1
                     }
                 },
-                PredefinedRecords = new List<string> { "5" }
+                //PredefinedRecords = new List<string> { "5" }
+                PredefinedRecord = new DataPredefinedFilterDefinition()
+                {
+                    PredefinedRecords = new List<string> { "5" }
+                }
             };
 
             var query = _dbContext.Posts